bau: Add a C# / dotnet core local matching service
For simplicity / parity with the ruby version this one uses dynamic types.
It would probably be more familiar if we added static types for the requests
and responses, but hey-ho.

The most painful bit of this is the "Hack" to overwrite the Content-Type in the
response so it doesn't contain "charset=utf-8". This definitely shouldn't be a
requirement for a local matching service (the MSA should be able to handle having
a charset in the header). We should confirm that there's no requirement to remove
charset, update the tests to allow it, and then remove the hack.

Authors: @richardtowers